# api/openapi.yaml
openapi: 3.2.0
info:
  title: Lighter API
  summary: Backend API for the Lighter weight loss tracker.
  version: 0.1.0
  description: |
    The contract between the Lighter frontend (`frontend/`) and backend (`backend/`).
    It covers everything the frontend currently keeps in local storage, plus the
    AI features that are mocked in `frontend/src/lib/ai.ts`.

    ## Conventions

    - **Auth.** Every operation needs a bearer token from `/auth/register` or
      `/auth/login`, unless it says otherwise. All data belongs to the signed-in
      user, so there is no user id in the paths.
    - **Units.** The API always uses metric units: weight in **kg**, length in **cm**,
      energy in **kcal**, macros in **grams**. `Profile.units` is a display
      preference only; converting to pounds and inches is the client's job.
    - **Dates and times.** Calendar days are `YYYY-MM-DD` (`format: date`) in the
      user's time zone (`Profile.timeZone`). Timestamps are RFC 3339 in UTC.
    - **One per day.** Weigh-ins, measurements, water and journal entries are
      identified by their date, so they are written with an idempotent
      `PUT /…/{date}` instead of `POST`.
    - **Lists.** Collections are wrapped as `{ "data": [...] }`. Large collections
      are paged with `limit` and an opaque `cursor`; follow `nextCursor` until it
      is `null`.
    - **Partial updates.** `PATCH` uses JSON Merge Patch
      (`application/merge-patch+json`, RFC 7396).
...
servers:
...
    name: local
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created and signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          description: Email or password is wrong. The response does not say which.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ---------------------------------------------------------------- Account
...
  /me/profile:
    get:
      tags: [Profile]
      operationId: getProfile
      summary: Get goals and preferences
      responses:
        "200":
          description: The user's profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: The user has not finished onboarding yet, so there is no profile.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    put:
      tags: [Profile]
      operationId: replaceProfile
      summary: Create or replace the profile
      description: Used when onboarding. Setting `startWeight` also records a weigh-in on `startDate`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Profile"
      responses:
        "200":
          description: Profile replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "201":
          description: Profile created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Profile]
      operationId: updateProfile
      summary: Change some goals or preferences
      description: |
        Changing `startWeight` updates the weigh-in on `startDate`.
        `macroGoals` is merged, so `{ "macroGoals": { "protein": 140 } }` changes only protein.
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/ProfilePatch"
            examples:
              calorieGoal:
                summary: Change the daily calorie goal
                dataValue:
                  calorieGoal: 1800
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------- Food
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token from `/auth/register` or `/auth/login`, sent as `Authorization: Bearer <token>`.'
...
  headers:
...
    RetryAfter:
      description: Seconds to wait before trying again.
      schema:
        type: integer
        minimum: 0
  responses:
...
    ValidationFailed:
      description: The request is malformed or a field is invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            invalidWeight:
              summary: A field failed validation
              dataValue:
                type: https://lighter.app/problems/validation-failed
                title: Your request is not valid.
                status: 400
                errors:
                  - pointer: /weight
                    detail: Must be greater than 0.
    Unauthorized:
      description: The bearer token is missing, invalid or expired.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Nothing exists at this URL for the signed-in user.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: The request conflicts with existing data, such as an email that is already registered.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    TooManyRequests:
      description: Too many requests. Wait for `Retry-After` seconds.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Id:
      type: string
      format: uuid
      examples: [0390357f-60db-4363-aae2-7b7fafd0d1e8]
    Date:
      type: string
      format: date
      description: A calendar day, `YYYY-MM-DD`.
      examples: ["2026-09-29"]
    Timestamp:
      type: string
      format: date-time
      examples: ["2026-09-29T07:42:10.512Z"]
...
    Grams:
      type: number
      minimum: 0
...
    Macros:
      type: object
      description: Protein, carbs and fat in grams.
      required: [protein, carbs, fat]
      properties:
        protein:
          $ref: "#/components/schemas/Grams"
        carbs:
          $ref: "#/components/schemas/Grams"
        fat:
          $ref: "#/components/schemas/Grams"
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        name:
          type: string
          minLength: 1
          maxLength: 80
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          writeOnly: true
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
    Session:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          description: Seconds until the token expires.
          examples: [3600]
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Id"
          readOnly: true
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
          maxLength: 80
        createdAt:
          $ref: "#/components/schemas/Timestamp"
          readOnly: true
...
    Profile:
      type: object
      additionalProperties: false
      required:
        [heightCm, startWeight, goalWeight, startDate, calorieGoal, macroGoals, trackMacros, waterGoal, glassMl, units, theme, timeZone]
      properties:
        heightCm:
          type: number
          exclusiveMinimum: 0
          maximum: 300
        startWeight:
          type: number
          exclusiveMinimum: 0
          description: Weight in kg on `startDate`. Stored as the weigh-in for that day.
        goalWeight:
          type: number
          exclusiveMinimum: 0
          description: Target weight in kg.
        startDate:
          $ref: "#/components/schemas/Date"
        calorieGoal:
          type: integer
          exclusiveMinimum: 0
          description: Daily net calorie target (eaten minus burned).
        macroGoals:
          $ref: "#/components/schemas/Macros"
        trackMacros:
          type: boolean
        waterGoal:
          type: integer
          minimum: 1
          maximum: 30
          description: Glasses per day.
        glassMl:
          type: integer
          minimum: 1
          maximum: 2000
          description: Size of one glass in millilitres.
        units:
          type: string
          enum: [metric, imperial]
          description: How the client should display values. The API always uses metric.
        theme:
          type: string
          enum: [system, light, dark]
        timeZone:
          type: string
          description: IANA time zone. Decides what "today" means for this user.
          examples: [Asia/Kolkata]
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
          readOnly: true
      examples:
        - heightCm: 175
          startWeight: 92.4
          goalWeight: 78
          startDate: "2026-06-01"
          calorieGoal: 1900
          macroGoals: { protein: 130, carbs: 190, fat: 65 }
          trackMacros: true
          waterGoal: 8
          glassMl: 250
          units: metric
          theme: system
          timeZone: Asia/Kolkata
    ProfilePatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        heightCm:
          $ref: "#/components/schemas/Profile/properties/heightCm"
        startWeight:
          $ref: "#/components/schemas/Profile/properties/startWeight"
        goalWeight:
          $ref: "#/components/schemas/Profile/properties/goalWeight"
        startDate:
          $ref: "#/components/schemas/Date"
        calorieGoal:
          $ref: "#/components/schemas/Profile/properties/calorieGoal"
        macroGoals:
          type: object
          additionalProperties: false
          minProperties: 1
          properties:
            protein:
              $ref: "#/components/schemas/Grams"
            carbs:
              $ref: "#/components/schemas/Grams"
            fat:
              $ref: "#/components/schemas/Grams"
        trackMacros:
          type: boolean
        waterGoal:
          $ref: "#/components/schemas/Profile/properties/waterGoal"
        glassMl:
          $ref: "#/components/schemas/Profile/properties/glassMl"
        units:
          $ref: "#/components/schemas/Profile/properties/units"
        theme:
          $ref: "#/components/schemas/Profile/properties/theme"
        timeZone:
          $ref: "#/components/schemas/Profile/properties/timeZone"

    # ---------- Food
...
    Problem:
      type: object
      description: An error, as defined by RFC 9457.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: Identifies the kind of error. Clients should branch on this, not on `title`.
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: array
          description: One item per invalid field, for validation errors.
          items:
            type: object
            required: [pointer, detail]
            properties:
              pointer:
                type: string
                description: JSON Pointer (RFC 6901) to the invalid field in the request body, or the parameter name.
                examples: [/weight]
              detail:
                type: string